Refactor Button and actions
