+++ packages/react/src/blocks/BigMessage/BigMessage.stories.mdx
@@ -13,6 +13,7 @@ import { themeArgType } from '../../lib'
     actions: { table: { disable: true } },
     viewportHeight: { table: { disable: true } },
     variant: { table: { disable: true } },
+    onAction: { action: 'action' },
   }}
 />
