+++ packages/react/src/views/View/View.tsx
@@ -1,11 +1,11 @@
 import { z } from 'zod'
+import { anyActionPayload } from '@fluentui/blocks-schemas'
 
 import {
   FluentPatternsProvider,
   themeName,
   ParseBoundary,
   withActionHandler,
-  anyActionPayload,
   translations,
   defaultTranslations,
 } from '../../lib'
