+++ packages/schemas/src/index.ts
@@ -1,1 +1,3 @@
 export * from './inlines'
+export * from './inputs'
+export * from './lib/actions'
