+++ packages/schemas/src/lib/actions.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod'
+
+export const actionPayload = z.object({
+  type: z.string(),
+  actionId: z.string(),
+})
+
+export const anyActionPayload = actionPayload.passthrough()
